PhishingEmailDetector/requirements.txt
pandas
scikit-learn
nltk
streamlit
joblib
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
